+++ lib/nfg_ui/bootstrap/components/modal.rb
@@ -25,7 +25,7 @@ module NfgUi
 
         def render
           super do
-            content_tag(:div, class: 'modal-dialog', role: 'document') do
+            content_tag(:div, class: modal_dialog_css_classes, role: 'document') do
               content_tag(:div, class: 'modal-content') do
                 capture do
                   concat(NfgUi::Bootstrap::Components::ModalHeader.new({ title: title }, view_context).render)
@@ -47,14 +47,36 @@ module NfgUi
           options.fetch(:render_in_body, true)
         end
 
+        def size
+          options.fetch(:size, nil)
+        end
+
         private
 
         def css_classes
-          super + ' fade'
+          [
+            component_css_class,
+            'fade'
+          ].join(' ')
+        end
+
+        def modal_dialog_css_classes
+          [
+            'modal-dialog',
+            (size_css_class if resized?)
+          ].join(' ').squish
         end
 
         def non_html_attribute_options
-          super.push((:footer if footer.present?), (:title if title.present?))
+          super.push((:footer if footer.present?), (:title if title.present?), :size)
+        end
+
+        def resized?
+          [:sm, :lg, :xl].include?(size)
+        end
+
+        def size_css_class
+          "#{component_css_class}-#{size}"
         end
 
         def assistive_html_attributes