+++ spec/lib/nfg_ui/components/patterns/modal_spec.rb
@@ -8,6 +8,7 @@ RSpec.describe NfgUi::Components::Patterns::Modal do
   it_behaves_like 'a component that includes the Describable utility module'
   it_behaves_like 'a component that includes the Renderable utility module'
   it_behaves_like 'a component that includes the Titleable utility module'
+  it_behaves_like 'a component that includes the Size trait module'
 
   it { expect(described_class.included_modules).to include NfgUi::Components::Utilities::Traitable }
 end